+++ app/lib/route_constraints/register_api_constraint.rb
@@ -2,10 +2,16 @@
 
 module RouteConstraints
   class RegisterApiConstraint
-    AVAILABLE_VERSIONS = ["v0.1", "v1.0-pre", "v1.0-rc"].freeze
+    AVAILABLE_VERSIONS = ["v0.1", "v1.0-rc"].freeze
 
     def self.matches?(request)
+      redirect_pre_release_requests_to_release_canidiate(request)
+
       AVAILABLE_VERSIONS.include?(request.path_parameters[:api_version])
+    end
+
+    def self.redirect_pre_release_requests_to_release_canidiate(request)
+      request.path_parameters[:api_version] = "v1.0-rc" if request.path_parameters[:api_version] == "v1.0-pre"
     end
   end
 end