+++ app/models/api/v1_0_pre/degree_attributes.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V10Pre
-    class DegreeAttributes < Api::V01::DegreeAttributes
-    end
-  end
-end