+++ app/models/api/v1_0_pre/hesa_trainee_detail_attributes.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V10Pre
-    class HesaTraineeDetailAttributes < Api::V01::HesaTraineeDetailAttributes
-      validates_with RulesValidator
-    end
-  end
-end