+++ app/models/api/v1_0_pre/placement_attributes.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V10Pre
-    class PlacementAttributes < Api::V01::PlacementAttributes
-    end
-  end
-end