+++ app/models/api/v1_0_pre/trainee_attributes.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V10Pre
-    class TraineeAttributes < Api::V01::TraineeAttributes
-    end
-  end
-end