+++ app/models/api/v1_0_pre/withdrawal_attributes.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V10Pre
-    class WithdrawalAttributes < Api::V01::WithdrawalAttributes
-    end
-  end
-end