+++ app/serializers/api/v1_0_pre/degree_serializer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Api
-  module V10Pre
-    class DegreeSerializer < Api::V01::DegreeSerializer
-    end
-  end
-end