+++ spec/requests/api/v1_0_pre/post_hesa_trainees_spec.rb
@@ -76,7 +76,7 @@ describe "`POST /api/v1.0-pre/trainees` endpoint" do
 
   context "when the request is valid" do
     before do
-      allow(Api::V10Pre::HesaMapper::Attributes).to receive(:call).and_call_original
+      allow(Api::V10Rc::HesaMapper::Attributes).to receive(:call).and_call_original
       allow(Trainees::MapFundingFromDttpEntityId).to receive(:call).and_call_original
     end