+++ spec/requests/api/v1_0_pre/trainees/put_degrees_spec.rb
@@ -200,7 +200,7 @@ describe "`PUT /trainees/:trainee_slug/degrees/:slug` endpoint" do
       let(:degrees_attributes) do
         non_uk_degree.attributes.symbolize_keys.slice(
           :country, :grade, :grade_uuid, :subject, :subject_uuid, :institution, :institution_uuid, :uk_degree, :uk_degree_uuid, :graduation_year, :non_uk_degree
-        ).merge(Api::V10Pre::DegreeSerializer.new(non_uk_degree).as_hash.symbolize_keys)
+        ).merge(Api::V10Rc::DegreeSerializer.new(non_uk_degree).as_hash.symbolize_keys)
       end
 
       before do
@@ -222,7 +222,7 @@ describe "`PUT /trainees/:trainee_slug/degrees/:slug` endpoint" do
             message: "This is a duplicate degree" },
         )
         expect(response.parsed_body["data"]).to contain_exactly(
-          JSON.parse(Api::V10Pre::DegreeSerializer.new(trainee.degrees.non_uk.first).as_hash.to_json),
+          JSON.parse(Api::V10Rc::DegreeSerializer.new(trainee.degrees.non_uk.first).as_hash.to_json),
         )
         expect {
           uk_degree.reload